# test-swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Test API",
    "version": "1.0.0",
    "description": "Test swagger configuration"
  },
  "paths": {
...
    "/api/appointments": {
      "post": {
        "summary": "Yangi appointment yaratish",
        "description": "Foydalanuvchi uchun yangi appointment yaratadi. Schedule ID orqali salon va employee ma'lumotlari avtomatik aniqlanadi.",
        "tags": [
          "Appointments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Appointment muvaffaqiyatli yaratildi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Appointment muvaffaqiyatli yaratildi"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Appointment"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Noto'g'ri ma'lumotlar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Avtorizatsiya xatosi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Schedule topilmadi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Barcha appointmentlarni olish",
        "description": "Admin va employee uchun barcha appointmentlarni ko'rish imkoniyati",
        "tags": [
          "Appointments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Appointmentlar ro'yxati",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Appointment"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Avtorizatsiya xatosi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/superadmin/login": {
      "post": {
        "summary": "Superadmin login",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "superadmin"
                  },
                  "password": {
                    "type": "string",
                    "example": "admin123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Muvaffaqiyatli login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Noto'g'ri so'rov"
          },
          "401": {
            "description": "Noto'g'ri ma'lumotlar"
          }
        }
      }
    },
    "/api/auth/admin/login": {
      "post": {
        "summary": "Admin login",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "admin_beauty_palace"
                  },
                  "password": {
                    "type": "string",
                    "example": "admin123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Muvaffaqiyatli login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Noto'g'ri so'rov"
          },
          "401": {
            "description": "Noto'g'ri ma'lumotlar"
          }
        }
      }
    },
...
    "/api/auth/employee/login": {
      "post": {
        "summary": "Employee login",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "johndoe"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Muvaffaqiyatli login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Username va password talab qilinadi"
          },
          "401": {
            "description": "Noto'g'ri username yoki password"
          },
          "500": {
            "description": "Server xatosi"
          }
        }
      }
    },
...
    "/api/payment-cards": {
      "post": {
        "summary": "Add a new payment card",
        "tags": [
          "Payment Cards"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPaymentCardRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment card added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/PaymentCard"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "get": {
        "summary": "Get user's payment cards",
        "tags": [
          "Payment Cards"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of user's payment cards",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PaymentCard"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/posts": {
      "get": {
        "summary": "Barcha postlarni olish",
        "tags": [
          "Posts"
        ],
        "responses": {
          "200": {
            "description": "Postlar ro'yxati muvaffaqiyatli olindi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server xatosi"
          }
        }
      },
      "post": {
        "summary": "Yangi post yaratish",
        "tags": [
          "Posts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/PostCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post muvaffaqiyatli yaratildi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Post muvaffaqiyatli yaratildi"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Post"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Noto'g'ri ma'lumotlar"
          },
          "401": {
            "description": "Autentifikatsiya talab qilinadi"
          },
          "403": {
            "description": "Admin huquqi talab qilinadi"
          },
          "500": {
            "description": "Server xatosi"
          }
        }
      }
    },
...
    "/api/user-salons": {
      "get": {
        "summary": "Barcha salonlarni ko'rish (faqat o'qish uchun)",
        "tags": [
          "User Salons"
        ],
        "security": [
          {
            "UserToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Salonlar ro'yxati muvaffaqiyatli qaytarildi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Salonlar ro'yxati"
                    },
                    "salons": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Salon"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 10
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Autentifikatsiya xatosi",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Token topilmadi, kirish rad etildi"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server xatosi"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PaymentCard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "card_holder_name": {
            "type": "string"
          },
          "expiry_month": {
            "type": "integer",
            "minimum": 1,
            "maximum": 12
          },
          "expiry_year": {
            "type": "integer"
          },
          "card_type": {
            "type": "string",
            "enum": [
              "visa",
              "mastercard",
              "uzcard",
              "humo",
              "unknown"
            ]
          },
          "phone_number": {
            "type": "string"
          },
          "is_default": {
            "type": "boolean"
          },
          "last_four_digits": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AddPaymentCardRequest": {
        "type": "object",
        "required": [
          "card_number",
          "card_holder_name",
          "expiry_month",
          "expiry_year",
          "phone_number"
        ],
        "properties": {
          "card_number": {
            "type": "string",
            "description": "16-19 digit card number"
          },
          "card_holder_name": {
            "type": "string"
          },
          "expiry_month": {
            "type": "integer",
            "minimum": 1,
            "maximum": 12
          },
          "expiry_year": {
            "type": "integer"
          },
          "phone_number": {
            "type": "string"
          },
          "is_default": {
            "type": "boolean",
            "default": false
          }
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Post ID"
          },
          "title": {
            "type": "string",
            "description": "Post sarlavhasi"
          },
          "description": {
            "type": "string",
            "description": "Post tavsifi"
          },
          "media_files": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Media fayllar ro'yxati"
          },
          "admin_id": {
            "type": "string",
            "format": "uuid",
            "description": "Admin ID"
          },
          "salon_id": {
            "type": "string",
            "format": "uuid",
            "description": "Salon ID"
          },
          "is_active": {
            "type": "boolean",
            "description": "Post faol holati"
          },
          "view_count": {
            "type": "integer",
            "description": "Ko'rishlar soni"
          },
          "like_count": {
            "type": "integer",
            "description": "Yoqtirishlar soni"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Yaratilgan vaqt"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Yangilangan vaqt"
          },
          "admin_username": {
            "type": "string",
            "description": "Admin username"
          },
          "admin_full_name": {
            "type": "string",
            "description": "Admin to'liq ismi"
          },
          "salon_name": {
            "type": "string",
            "description": "Salon nomi"
          }
        }
      },
      "PostCreate": {
        "type": "object",
        "required": [
          "title",
          "description"
        ],
        "properties": {
          "title": {
            "type": "string",
            "description": "Post sarlavhasi",
            "example": "Yangi xizmat"
          },
          "description": {
            "type": "string",
            "description": "Post tavsifi",
            "example": "Bizda yangi soch turmagi xizmati mavjud"
          },
          "media_files": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "binary"
            },
            "description": "Yuklash uchun media fayllar"
          }
        }
      },
...
    },
    "securitySchemes": {
      "UserToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "User uchun maxsus yaratilgan read-only token"
      }
    }
  },
...
}